v0.5a osmplesher;
огромное количество изменений.

- :musical_note: музыкальные функции перенесены в music.py
- :musical_note: play => rock, choice >> play
- :musical_note: choice выводит embed сообщение (красиво :heart_eyes: )

- :paperclip: blankfix теперь (логично) так и называется

- :heavy_exclamation_mark: some changes in evala

- :computer: комментарии++

-: bug: куча багов в подарок :kiss:
